Rename `UnassignedMatchDto` to `CompactMatchDto` and extend it
Previously, `UnassignedMatchDto` ignored some fields which could
have been deserialized from the response received from `GET
/api/matches?matchIds=`. This endpoint's response's usefulness goes way beyond
simply fetching unassigned matches, so it makes sense to actually
deserialize all available fields, and rename the dto object to
`CompactMatchDto`.
